Restored some basic functionality lost in merge

+++ GroundStation/GroundStation/Panels/DropStatus.cs
@@ -21,7 +21,7 @@ namespace GroundStation.Panels
         //Requires: altitude in feet and time in seconds
         //Modifies: Nothing
         //Effects: drop status display
-        public void UpdateDrop(int alt_ft, int time_s)
+        public void UpdateDrop(double alt_ft, double time_s)
         {
             Payload_Time_Status.Text = "Time: " + time_s + " s"; //States time that payload was dropped.
             Payload_Alt_Status.Text = "Altitude: " + alt_ft + " ft"; //States altitude that payload was dropped.
